chenge type count

+++ backend/src/main/java/ru/nsu/sber_portal/ccfit/services/OrderCheckServiceUtility.java
@@ -19,7 +19,7 @@ import static java.util.Optional.ofNullable;
 @Slf4j
 public class OrderCheckServiceUtility {
 
-    protected static final Integer EMPTY_ORDER = 0;
+    protected static final Long EMPTY_ORDER = 0L;
 
     protected static final Long EMPTY_PRICE = 0L;
